pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Sergei Zuev <zuevsn86@gmail.com>"]
readme = "README.md"
packages = [
  { include = "page_analyzer" },
]

[tool.poetry.dependencies]
python = "^3.8.10"
flask = "^3.0.3"
gunicorn = "^23.0.0"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
